Fixes #10: sets expiry (double retry delay) on keys
Sets expiry (double retry delay time just to be safe) on failure keys in
MultipleWithRetrySuppression failure backend. If retry_delay is not
set, meaning we bypass the scheduler and enqueue the job again
immediately, we don't even bother writing the failure key.

This also removes the broken attempt at cleaning up these keys, which
doesn't work because the mixin never happens due to Resque's forking
model, and because the job can be run on a different box / worker
process, and because the job might not be actually run again (it could
be cleared etc.)