ensure we make calls to Redis only once for logging
The log message in question needs to distinguish between "job is not
retriable" and "job IS retriable, but isn't actually being retried right
now".

We also don't want to go to Redis more than is required, so intermediate
variables to the rescue.  Also happens to straighten out that
double-negative, DeMorgan's Law-requiring `if` statement expression.