Update README.md
Improved README.md to explain the steps needed to start the 'resque-web' server interface properly with the 'resque-retry' features. 

The problems as I identified them essentially were: 
A. The README proposed a rack configuration, but didn't include an example command for a user to start the server. Solution: Added a simple example on how to start the server with bundler by loading the proposed rack configuration. 
B. The README didn't include the alternate option of starting the resque-web interface using the built-in 'resque-web' command. Solution: Included a simple 3 step explanation on how to use the resque-web command by sending it a configuration file as a parameter