Don't validate ignored exceptions when retry is included on job
Ignored and fatal exceptions are different functions. Fatal exceptions cause
the job to not run through the rest of the retry logic at all. Ignored exceptions
cause the job retry counter to not be increased.

Setting @fatal_exceptions causes custom custom criteria callbacks to be skipped.
This makes it impossible to ignore exceptions and implement custom retry criteria.